feat(core): extract git/hu-snapshot to @karajan/core (KJC-TSK-0511)
Verbatim copy of src/git/hu-snapshot.js to packages/core/src/hu-snapshot.js
(only the runCommand import path changes: ../utils/process -> ./process).
Original src/git/hu-snapshot.js becomes a 9-line re-export shim so CLI
callers keep working.

packages/hu-board/src/plan-mutations.js no longer reaches into
../../../src for the snapshot restore — the dynamic import now targets
@karajan/core/hu-snapshot.

After this PR there is exactly one remaining static + dynamic import from
hu-board into src/: brain/standby-scheduler.js (which itself pulls
standby-store.js). PR8 will close that.

Tests: 44/44 core (5 new for hu-snapshot subpath), 14/14 CLI shim,
49/49 hu-board plan-mutations.

Net delta: +52 LOC (159 add / 107 del).
