+++ packages/core/src/index.js
@@ -13,3 +13,4 @@ export * from "./plan-id.js";
 export * from "./plan-hu-ops.js";
 export * from "./plan-validation.js";
 export * from "./process.js";
+export * from "./hu-snapshot.js";
